reduce trampoline code, correct warnings

+++ main.c
@@ -8,7 +8,7 @@
 #include <conio.h>
 
 DECLARE_BANKED(setupScreen, BANK_1, void far_setupScreen(), trampoline())
-DECLARE_BANKED(countup, BANK_1, int far_countup(int a, int b, int c), return trampoline(a, b, c))
+DECLARE_BANKED(countup, BANK_1, int far_countup(int a, int b, int c), return (int) trampoline(a, b, c))
 DECLARE_BANKED(cputs, BANK_1, void far_cputs(const char* s), trampoline(s))
 DECLARE_BANKED(checkresult, BANK_1, void far_checkresult(int s), trampoline(s))
